Properly target Java 8
Java 9+ added an overload of various methods in ByteBuffer. When compiling with Java 11+ and targeting Java 8 bytecode
the resulting signatures are invalid for JDK 8, so accesses below result in NoSuchMethodError. Accessing the
methods through the interface class works around the problem.

Ideally, we'd use "javac --release 8", but that doesn't work because Unsafe is not available in the signature data
for that profile.